Notebook documentation improvements
